web-ide/backend/runtime_python/requirements.txt
fastapi
uvicorn
pymongo
psycopg2-binary
jupyter-client
requests
pydantic
python-dotenv
motor